+++ forge/src/legacy-1.16.5/java/com/iamkaf/konfig/api/v1/KonfigClientScreens.java
@@ -14,4 +14,8 @@ public final class KonfigClientScreens {
     public static Screen create(String modId, Screen parent) {
         return new KonfigConfigScreen(parent, modId);
     }
+
+    public static Screen create(String modId, String title, Screen parent) {
+        return new KonfigConfigScreen(parent, modId, title);
+    }
 }